refactor: extract playground command error helpers
